Renamed ApiControllerBase to ApiController Extracted ApiController list endpoint into readonly base class
